+++ finagle-mysql/src/test/scala/com/twitter/finagle/mysql/HandshakeSettingsTest.scala
@@ -1,7 +1,11 @@
 package com.twitter.finagle.mysql
 
 import com.twitter.finagle.Stack
-import com.twitter.finagle.mysql.param.{Charset, Credentials, Database, FoundRows}
+import com.twitter.finagle.mysql.param.Charset
+import com.twitter.finagle.mysql.param.Credentials
+import com.twitter.finagle.mysql.param.Database
+import com.twitter.finagle.mysql.param.Interactive
+import com.twitter.finagle.mysql.param.FoundRows
 import org.scalatest.funsuite.AnyFunSuite
 
 class HandshakeSettingsTest extends AnyFunSuite {
@@ -11,6 +15,7 @@ class HandshakeSettingsTest extends AnyFunSuite {
     Capability.MultiResults
   )
   private val withFoundRows = initial + Capability.FoundRows
+  private val withInteractive = initial + Capability.Interactive
 
   test("HandshakeSettings adds FoundRows by default") {
     val settings = HandshakeSettings(clientCapabilities = initial)
@@ -19,6 +24,16 @@ class HandshakeSettingsTest extends AnyFunSuite {
 
   test("HandshakeSettings returns initial when found rows is disabled") {
     val settings = HandshakeSettings(clientCapabilities = initial, enableFoundRows = false)
+    assert(settings.calculatedClientCapabilities == initial)
+  }
+
+  test("HandshakeSettings adds Interactive by default") {
+    val settings = HandshakeSettings(clientCapabilities = initial)
+    assert(settings.calculatedClientCapabilities == withInteractive)
+  }
+
+  test("HandshakeSettings returns initial when interactive is disabled") {
+    val settings = HandshakeSettings(clientCapabilities = initial, interactive = false)
     assert(settings.calculatedClientCapabilities == initial)
   }
 
@@ -40,13 +55,15 @@ class HandshakeSettingsTest extends AnyFunSuite {
       Charset(MysqlCharset.Binary) +
       Credentials(Some("user123"), Some("pass123")) +
       Database(Some("test")) +
-      FoundRows(false)
+      FoundRows(false) +
+      Interactive(false)
     val settings = HandshakeSettings(params)
     assert(settings.username == Some("user123"))
     assert(settings.password == Some("pass123"))
     assert(settings.database == Some("test"))
     assert(settings.charset == MysqlCharset.Binary)
     assert(!settings.enableFoundRows)
+    assert(!settings.interactive)
   }
 
 }